# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /canvas-projects/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [canvas]
      summary: 读取当前用户画布内容和 revision
      responses:
        "200":
          description: 信封 data.project 包含完整画布及服务端 revision
        "404":
          description: 画布不存在或不属于当前用户
    put:
      tags: [canvas]
      summary: 按 revision 保存画布内容
      description: 新建传 revision 0，更新传读取到的版本；成功原子加 1。忽略客户端 viewport 和 remoteContentHash，updatedAt 由服务端生成。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project]
              properties:
                project:
                  type: object
                  required: [id, revision]
                  properties:
                    id: { type: string }
                    revision: { type: integer, format: int64, minimum: 0 }
                  additionalProperties: true
      responses:
        "200":
          description: 信封 data.project 返回摘要及递增后的 revision
        "400":
          description: 请求、版本或素材引用不合法
        "409":
          description: 版本或资源状态冲突，原内容未被覆盖；保留草稿后加载最新版
        "428":
          description: 缺少 revision，需刷新客户端并重新读取
  /canvas-projects/{id}/history:
    get:
      tags: [canvas]
      summary: 读取当前画布版本与最近 20 份历史摘要
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 信封 data 包含 currentRevision 和 snapshots；每项含 id、revision、title、nodeCount、connectionCount、payloadBytes、reason、contentUpdatedAt、createdAt，不含正文
        "404":
          description: 画布不存在或无权访问
  /canvas-projects/{id}/history/{snapshotId}:
    get:
      tags: [canvas]
      summary: 按需读取历史正文
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: snapshotId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 信封 data 包含 snapshot 摘要和 project 正文
        "404":
          description: 画布或快照不存在、已过期或无权访问
  /canvas-projects/{id}/history/{snapshotId}/restore:
    post:
      tags: [canvas]
      summary: 备份当前内容并将历史内容恢复为新版本
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: snapshotId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              properties:
                revision: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: 信封 data.project 返回恢复后的摘要及新 revision；当前 ID、归属与项目关联保留
        "404":
          description: 画布或快照不存在、已过期或无权访问
        "409":
          description: 当前版本或素材状态已变化，恢复未执行
        "428":
          description: 缺少当前 revision
...
  /resources:
    get:
      tags: [resources]
      summary: 当前用户资源列表
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
    post:
      tags: [resources]
      summary: 上传资源文件
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "403":
          description: 配额用尽（code=40301, reason=quota_exceeded）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /resources/access:
    post:
      tags: [resources]
      summary: 批量签发当前用户资源访问描述
      description: 统一处理展示、下载、复制、浏览器字节处理和模型输入；客户端不能通过查询参数选择 direct/proxy。每项独立成功或失败。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 100
              items:
                type: object
                required: [resourceId]
                properties:
                  resourceId: {type: string}
                  purpose: {type: string, enum: [display, copy, download, browser-process, provider-input], default: display}
                  variant: {type: string, enum: [original, playback], default: original}
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /resources/{id}/file:
    get:
      tags: [resources]
      summary: 读取资源二进制（非 JSON 信封）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: access
          schema: { type: string, enum: ["1"] }
          description: 仅返回统一访问描述，不返回文件正文。
        - in: query
          name: variant
          schema: { type: string, enum: [original, playback], default: original }
        - in: query
          name: purpose
          schema: { type: string, enum: [display, download, browser-process], default: display }
      responses:
        "200":
          description: 文件流
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /assets:
    get:
      tags: [assets]
      summary: 用户素材库
      parameters:
        - in: query
          name: includeEntities
          schema: { type: integer, enum: [0, 1], default: 0 }
          description: 画布资产入口设为 1 时，同时返回角色卡实体；默认仍只返回媒体与文本素材
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
  /assets/batch-delete:
    post:
      tags: [assets]
      summary: 批量彻底删除当前用户素材
      description: 顶层 JSON 数组，一次请求、一个数据库事务。所有 ID 必须属于当前用户，不受任务、画布或历史引用阻塞；共享文件保留，其余通过 Outbox 异步清理，失败重试。任一素材不存在、越权或数据库写入失败时整批不删除。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 1000
              items:
                type: string
                minLength: 1
            example: ["asset-1", "asset-2"]
      responses:
        "200":
          description: 成功信封，data.ids 返回本次请求的素材 ID 数组；物理文件清理由后台执行。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [ids]
                        properties:
                          ids:
                            type: array
                            items: { type: string }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /assets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    delete:
      tags: [assets]
      summary: 彻底删除当前用户素材
      description: 不依赖素材 archived 状态，也不受画布、任务、项目或历史快照引用阻塞；其他素材共享的物理对象仍保留。
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - upstream_dns_failed
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
...
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"